adds topics with drop down for stances
There are now topics shown on the home screen. They currently don't link to anything but when one of them is pressed there is a drop down to choose your stance, if you're for or against it. (Can probably be changed to more of a popup to choose stance if needed)