Finished UI for profile page
It isn't editable yet but the profile page now shows the users picture, name, and email address. There are sections for a user score that we talked about a while ago and also a section for how many chats a person has had. There is an about section that is not editable yet either.